% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_by_labels.R
\name{cluster_by_labels}
\alias{cluster_by_labels}
\title{cluster_by_labels}
\usage{
cluster_by_labels(data, elements, labels, return_labels = TRUE)
}
\arguments{
\item{data}{Data frame of data elements and associated labels.}

\item{elements}{Column name for element id.}

\item{labels}{Column name for labels.}

\item{return_labels}{Return label columns for elements and clusters.}
}
\description{
This function allows you to input a data frame of elements and associated labels and clusters the elements
based on whether they share labels with other elements. Returns a data frame of the supplied element IDs and
assigned cluster IDs.
}
\keyword{cluster}
\keyword{components}
\keyword{connected}
\keyword{graph}
\keyword{labels}
\keyword{network}
